requirements.txt
praw>=7.7.0
rich>=13.7.0
requests>=2.31.0
python-dotenv>=1.0.0
Flask>=2.0.0
flask-smorest>=0.35.0